New HealthCheck package - Moved HealthCheckMixin.py
